deploy: ask the OCI runtime to set RLIMIT_NOFILE to 524288
The default definition created by containerd asks the OCI runtime to
set RLIMIT_NOFILE to 1024. The virtio-fs device/server integrated in
libkrun needs a higher limit, so modify the spec to set RLIMIT_NOFILE
to a safe value (524288).

This is a temporary workaround. In the future, crun will pass the
rlimit configuration set in the spec file to libkrun (to be set inside
the VM), and use a proper configuration for the container context
where the VMM is executed.

Signed-off-by: Sergio Lopez <slp@redhat.com>